+++ .gitignore
@@ -32,3 +32,5 @@ Thumbs.db
 #.idea/workspace.xml - remove # and delete .idea if it better suit your needs.
 .gradle
 build/
+
+lint-results.sarif
