--- android-studio-2/build.gradle
@@ -1,44 +1,0 @@
-/*
- * Copyright (C) 2015 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply plugin: 'java'
-
-repositories {
-    google()
-    jcenter()
-}
-
-ext {
-    lintVersion = '26.0.0-beta2'
-}
-
-dependencies {
-    // For a description of the below dependencies, see the main project README
-    compile "com.android.tools.lint:lint-api:$lintVersion"
-    compile "com.android.tools.lint:lint-checks:$lintVersion"
-    testCompile "junit:junit:4.11"
-    testCompile "com.android.tools.lint:lint:$lintVersion"
-    testCompile "com.android.tools.lint:lint-tests:$lintVersion"
-    testCompile "com.android.tools:testutils:$lintVersion"
-}
-
-jar {
-    manifest {
-        attributes("Lint-Registry": "com.example.google.lint.MyIssueRegistry")
-    }
-}
-
-defaultTasks 'assemble'
