--- android-studio-2/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Wed Aug 09 14:33:29 PDT 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.1-rc-1-all.zip
