--- android-studio-2/src/main/java/com/example/google/lint/MyIssueRegistry.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (C) 2015 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.google.lint;
-
-import com.android.tools.lint.client.api.IssueRegistry;
-import com.android.tools.lint.detector.api.Issue;
-
-import java.util.Collections;
-import java.util.List;
-
-/**
- * The list of issues that will be checked when running <code>lint</code>.
- */
-@SuppressWarnings("unused")
-public class MyIssueRegistry extends IssueRegistry {
-    @Override
-    public List<Issue> getIssues() {
-        return Collections.singletonList(MainActivityDetector.ISSUE);
-    }
-}
