--- android-studio-3/build.gradle
@@ -1,24 +1,0 @@
-buildscript {
-    ext {
-        lintVersion = '26.5.0'
-    }
-
-    repositories {
-        google()
-        jcenter()
-    }
-    dependencies {
-        classpath "com.android.tools.build:gradle:3.5.2"
-    }
-}
-
-allprojects {
-    repositories {
-        google()
-        jcenter()
-    }
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
-}
