--- android-studio-3/checks/build.gradle
@@ -1,22 +1,0 @@
-apply plugin: 'java-library'
-
-dependencies {
-    // For a description of the below dependencies, see the main project README
-    compileOnly "com.android.tools.lint:lint-api:$lintVersion"
-    compileOnly "com.android.tools.lint:lint-checks:$lintVersion"
-    testImplementation "junit:junit:4.12"
-    testImplementation "com.android.tools.lint:lint:$lintVersion"
-    testImplementation "com.android.tools.lint:lint-tests:$lintVersion"
-    testImplementation "com.android.tools:testutils:$lintVersion"
-}
-
-sourceCompatibility = "1.8"
-targetCompatibility = "1.8"
-
-jar {
-    manifest {
-        // Only use the "-v2" key here if your checks have been updated to the
-        // new 3.0 APIs (including UAST)
-        attributes("Lint-Registry-v2": "com.example.lint.checks.SampleIssueRegistry")
-    }
-}
