--- android-studio-3/checks/src/main/java/com/example/lint/checks/SampleIssueRegistry.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (C) 2017 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.lint.checks;
-
-import com.android.tools.lint.client.api.IssueRegistry;
-import com.android.tools.lint.detector.api.ApiKt;
-import com.android.tools.lint.detector.api.Issue;
-
-import java.util.Collections;
-import java.util.List;
-
-/*
- * The list of issues that will be checked when running <code>lint</code>.
- */
-public class SampleIssueRegistry extends IssueRegistry {
-    @Override
-    public List<Issue> getIssues() {
-        return Collections.singletonList(SampleCodeDetector.ISSUE);
-    }
-
-    @Override
-    public int getApi() {
-        return ApiKt.CURRENT_API;
-    }
-}
-
