--- android-studio-3/library/build.gradle
@@ -1,18 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 29
-    defaultConfig {
-        minSdkVersion 15
-        targetSdkVersion 29
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-}
-
-/** Package the given lint checks library into this AAR  */
-dependencies {
-    lintChecks project(':checks')
-}
