--- android-studio-3/settings.gradle
@@ -1,1 +1,0 @@
-include ':checks', ':library'
