--- android-studio-4/build.gradle
@@ -1,27 +1,0 @@
-buildscript {
-    ext {
-        gradlePluginVersion = '4.1.0-alpha02'
-        lintVersion = '27.1.0-alpha02'
-        kotlinVersion = '1.3.70'
-    }
-
-    repositories {
-        google()
-        jcenter()
-    }
-    dependencies {
-        classpath "com.android.tools.build:gradle:$gradlePluginVersion"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
-    }
-}
-
-allprojects {
-    repositories {
-        google()
-        jcenter()
-    }
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
-}
