--- android-studio-4/checks/build.gradle
@@ -1,25 +1,0 @@
-apply plugin: 'java-library'
-apply plugin: 'kotlin'
-apply plugin: 'com.android.lint'
-
-lintOptions {
-    htmlReport true
-    htmlOutput file("lint-report.html")
-    textReport true
-    absolutePaths false
-    ignoreTestSources true
-}
-
-dependencies {
-    // For a description of the below dependencies, see the main project README
-    compileOnly "com.android.tools.lint:lint-api:$lintVersion"
-    compileOnly "com.android.tools.lint:lint-checks:$lintVersion"
-    compileOnly "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
-    testImplementation "junit:junit:4.13"
-    testImplementation "com.android.tools.lint:lint:$lintVersion"
-    testImplementation "com.android.tools.lint:lint-tests:$lintVersion"
-    testImplementation "com.android.tools:testutils:$lintVersion"
-}
-
-sourceCompatibility = "1.8"
-targetCompatibility = "1.8"
