--- android-studio-4/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Sep 01 06:52:38 PDT 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.2.1-all.zip
