--- android-studio-4/library/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<manifest package="com.example.lint.library">
-
-</manifest>
