--- android-studio-4/settings.gradle
@@ -1,1 +1,0 @@
-include ':checks', ':library'
