+++ app/src/main/java/com/android/example/Test.kt
@@ -1,0 +1,10 @@
+package com.android.example
+
+class Test {
+    // We have a custom lint check bundled with :library
+    // that this module depends on. The lint check looks
+    // for mentions of "lint", which should trigger an
+    // error
+    val s = "lint"
+    fun lint() { }
+}
