+++ checks/src/main/resources/META-INF/services/com.android.tools.lint.client.api.IssueRegistry
